+++ front/src/store/modules/signInSlice.js
@@ -1,6 +1,4 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { signIn } from '../../api/signInAPI';
-import axios from '../../api/axios';
+import { createSlice } from '@reduxjs/toolkit';
 import { validationCheck } from '../../util/util';
 
 const initialState = {
@@ -9,26 +7,7 @@ const initialState = {
   inputValue: { email: '', password: '' },
   inputStatus: { email: '', password: '' },
   inputHelperText: { email: '', password: '' },
-  isLogin: false,
-  firstLogin: false,
 };
-
-export const signInAsync = createAsyncThunk(
-  'signIn/getTokens',
-  async (params, thunkAPI) => {
-    try {
-      const response = await signIn(params);
-      axios.defaults.headers.common['Authorization'] =
-        'Bearer ' + response.headers.authorization;
-
-      return response;
-    } catch (error) {
-      return thunkAPI.rejectWithValue({
-        errorMessage: error.response.data.message,
-      });
-    }
-  }
-);
 
 export const signInSlice = createSlice({
   name: 'signIn',
@@ -43,40 +22,6 @@ export const signInSlice = createSlice({
         state.inputHelperText[ele.id] = ele.inputHelperText;
       }
     },
-  },
-  extraReducers: (builder) => {
-    builder
-      .addCase(signInAsync.pending, (state) => {
-        state.loading = true;
-        state.error = null;
-        state.inputValue = { email: '', password: '' };
-        state.inputStatus = { email: '', password: '' };
-        state.inputHelperText = { email: '', password: '' };
-        state.isLogin = false;
-        state.firstLogin = false;
-      })
-      .addCase(signInAsync.fulfilled, (state, action) => {
-        state.loading = false;
-        state.error = null;
-        state.inputValue = { email: '', password: '' };
-        state.inputStatus = { email: '', password: '' };
-        state.inputHelperText = { email: '', password: '' };
-        state.isLogin = true;
-        state.firstLogin = action.payload.data.firstLogin;
-      })
-      .addCase(signInAsync.rejected, (state, action) => {
-        state.loading = false;
-        if (action.payload) {
-          state.error = action.payload.errorMessage;
-        } else {
-          state.error = action.error;
-        }
-        // state.inputValue = { email: '', password: '' };
-        // state.inputStatus = { email: '', password: '' };
-        // state.inputHelperText = { email: '', password: '' };
-        state.isLogin = false;
-        state.firstLogin = action.payload.data.firstLogin;
-      });
   },
 });
